+++ src/pools/cluster/fixed.ts
@@ -101,6 +101,11 @@ export class FixedClusterPool<
   }
 
   /** @inheritDoc */
+  public get size (): number {
+    return this.numberOfWorkers
+  }
+
+  /** @inheritDoc */
   protected get full (): boolean {
     return this.workerNodes.length === this.numberOfWorkers
   }